+++ packages/aesthetic-adapter-typestyle/index.js
@@ -1,0 +1,7 @@
+/**
+ * @copyright   2017, Miles Johnson
+ * @license     https://opensource.org/licenses/MIT
+ */
+
+// eslint-disable-next-line
+module.exports = require('./lib/NativeAdapter').default;